+++ airflow/migrations/versions/0053_1_10_10_add_rendered_task_instance_fields_table.py
@@ -15,13 +15,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""Add ``RenderedTaskInstanceFields`` table
+"""Add ``RenderedTaskInstanceFields`` table.
 
 Revision ID: 852ae6c715af
 Revises: a4c2fd67d16b
 Create Date: 2020-03-10 22:19:18.034961
 
 """
+
 from __future__ import annotations
 
 import sqlalchemy as sa
@@ -41,7 +42,7 @@ TABLE_NAME = "rendered_task_instance_fields"
 
 
 def upgrade():
-    """Apply Add ``RenderedTaskInstanceFields`` table"""
+    """Apply Add ``RenderedTaskInstanceFields`` table."""
     json_type = sa.JSON
     conn = op.get_bind()
 
@@ -64,5 +65,5 @@ def upgrade():
 
 
 def downgrade():
-    """Drop RenderedTaskInstanceFields table"""
+    """Drop RenderedTaskInstanceFields table."""
     op.drop_table(TABLE_NAME)