+++ airflow/migrations/versions/0085_2_1_3_add_queued_at_column_to_dagrun_table.py
@@ -15,13 +15,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""Add ``queued_at`` column in ``dag_run`` table
+"""Add ``queued_at`` column in ``dag_run`` table.
 
 Revision ID: 97cdd93827b8
 Revises: a13f7613ad25
 Create Date: 2021-06-29 21:53:48.059438
 
 """
+
 from __future__ import annotations
 
 import sqlalchemy as sa
@@ -38,11 +39,11 @@ airflow_version = "2.1.3"
 
 
 def upgrade():
-    """Apply Add ``queued_at`` column in ``dag_run`` table"""
+    """Apply Add ``queued_at`` column in ``dag_run`` table."""
     op.add_column("dag_run", sa.Column("queued_at", TIMESTAMP, nullable=True))
 
 
 def downgrade():
-    """Unapply Add ``queued_at`` column in ``dag_run`` table"""
+    """Unapply Add ``queued_at`` column in ``dag_run`` table."""
     with op.batch_alter_table("dag_run") as batch_op:
         batch_op.drop_column("queued_at")